Trying to avoid overwriting the shared memory
Suppose S1 is linked in a composition to S2 in the same JVM, and S1
poolsize is greater than 1. The shared memory keys of S1 are named S1X,
where X is a unique ID given by the DPE.

When doing multicore requests, the same engine S1N can be executed
twice before the result of the first execution is received by S2.
Suppose the results of S1N are "A" and "B".

In this situation, S2 will receive the two messages, and it will get the
data from the shared memory. But both times it will get "B", because the
shared memory for the engine S1N was overwritten by the consecutive
requests.

Solve this by ensuring that every result of S1N is stored with its own
key. Do it by constructing the shared memory keys as S1X:Y, where Y is
the id of the request.