DPE: fix the IP address used
Use the internal xMsgUtilty.host_to_ip  method to get the IP address for
the DPE name, because the CUtility.getIPAddress helper method is not
returning the same when the host has multiple addresses.

This fix was missing on 9d1956fadf1837a0b8b3b109629cd6b4ba3d870b.