Service: register after creating engines and subscription
Don't register with the xMsg pub/sub registration before starting the
subscription. The service may appear as ready (i.e. registered), but the
first requests will be lost because the subscription is still being
created. Also, some engines objects may still being created, but the
first created engine already registered the service.

In order to ensure the subscription is running, the registration has to
be sent after the ServiceDispatcher is created. This also ensures that
all the engine objects are already created.