ServiceEngine: refactor receiving data
Don't pass the data and the metadata, but pass the message and extract
the data inside.

Clara handles serialization and shared memory now, so the data may need
serialization, or it needs to be fetched from the shared memory. It is
better to pass the message and let the engine handle the data, instead
of doing it in the service callback.