Use a semaphore to acquire an Engine for execution
Using check-then-act and an AtomicBoolean may result in a race condition.
The service could be seen as available by two or more parallel requests,
because the service will be put as unavailable only when process() is
called. Thus, if the task to run process() is still on queue when
another request arrives, the service will appear as available.

With a semaphore, the engine will be locked as soon as it is acquired,
preventing multiple tasks using the same engine.