ServiceEngine: report the last execution time
Set the last execution time in the output EngineData, not the average.

Use a field to save the time, so it can be accessed by the
updateMetadata method (it is better to set all the output metadata
fields on a single method).

Having the last execution time as a field is not an issue, since the
ServiceEngine will always run sequentially like in a single thread.