DPE: accept front-end address
Remove RegHost, RegPort, CloudProxyHost and CloudProxyPort from
command line options. Only accept feHost and fePort, to pass the
location of the front-end node.

The new Clara topology consist of normal DPEs and a single front-end DPE
with public-access, that also acts as an optional gateway in case all
DPEs are part of an internal network.