BaseOrchestrator: fix listen/unlisten thread safety
Use a concurrent map to handle concurrent requests, but with lazy
subscription wrappers as the objects instead of the xMsg subscription
handlers.

The wrappers prevent creating multiple subscriptions if two threads call
listen at the same time and a subscription to the topic has not being
created yet, because an object needs to be created in order to use the
putIfAbsent method.

If the xMsg subscription handlers were used, two subscriptions would be
created, but only one would be put in the map and the other would need
to be discarded, which is complicated because unsubscribe can throw but
the proper exception should be related to the duplicated subscription,
not a failure unsubscribing.

If the wrapper was put in the map, then no other thread can overwrite it
with a different wrapper, and it is safe to start the subscription.

Another threads may concurrent call unlisten when the subscription has
not started yet (why?), so the xMsg handler in the wrapper must be
checked for nullity before calling unsubscribe and removing the wrapper.