dpe: change detection of FE or worker on command-line
If "fe_host" and "fe_port" are not present in command line arguments,
then the DPE is a front-end. This is the default. If the front-end host
(and port) are given as arguments, then the DPE is a worker that
registers with a remote front-end.

But the "fe_host" and "fe_port" could be passed along with the "fe"
switch in order to use the DPE as front-end and set a front-end address
with a different value than the local address. This is confusing. If the
DPE is a front-end then the same address should be used for the
registrar (front-end) and the proxy.

To avoid complications, ignore the "fe" switch from the command-line
to ensure that if the "fe_host" and "fe_port" are used only with
workers DPEs. If the DPE is a front-end, then the local address can be
customized in the same way as a worker DPE.