base: change signature of abstract start() method
Use the start method to start the Clara component in a thread-safe way,
i.e., the subscritions should not created on the constructor because the
this reference may escape before the component is constructed.

Do not pass a component as an argument to start. The component is
already given in the constructor.