dpe: fix request to shutdown
The xMsg actor must be closed, or the process will hang.

Also, any command that runs inside the callback to stop the actor must
run in a different thread, otherwise things will go wrong (destroying
the actor waits for all callbacks tasks to stop, but destroying was
running as a callback task).