report: calculate the number of container requests
Set the number of requests of the container equal to the sum of the
requests of all its services.

It is better to have the numbers in the report matching each other.
If the number of requests of the container were updated separately,
the sum of the requests of its services would probably be different
(while the report is being generated, some other thread could update
some of the values).

Anyway, it is much more simple just to sum the requests than having a
field in the container report to be updated any time a service request
is received.