sys: start service engines in parallel
When deploying a service, the service will create as many engine objects
as the defined pool size.

Each of them contains its own xMsg actor, and its start method will
cache a connection to the local proxy. When xMsg creates a connection,
it validates that the socket is connected by waiting for a control
message response from the proxy.

Thus, starting all engines in a serial loop takes a lot of time, because
each one will create a local connection.

Use the parallel stream API from Java 8 to speed up this loop.