sys: stop children components in parallel
Stopping a CLARA component requires (not in this exact order):

- Close any resource used by the component.
- Deregister the component from the front-end.
- Stop subscriptions and callbacks (destroy the xMsg actor).
- Stop all children components.

Stopping all the children in a serial loop may take a lot of time,
and shutting down the DPE becomes a problem for users.

Use the parallel stream API from Java 8 to speed up the loop that stops
the children components.