build.gradle: cache snapshots for just 30 minutes
The default refresh time for changing modules is 24 hours. But since
xMsg is being developed in parallel with CLARA, the CI server may need
the latest xMsg snapshot, and waiting 24 hours is too much.

A window of 30 minutes seems good enough. It prevents checking for new
snapshots at every build, while getting a new snapshot almost as soon as
published.

It is important to ensure that the latest xMsg snapshot is in the Maven
repository (by at least half an hour) before pushing the CLARA changes
that use the updated xMsg version, or the CI build will fail.