dpe: use a better request string for JSON report
The DPE returns the full registration and runtime data, so using a
request string that says "registration only" is not accurate.
Also, it is more simple for the orchestrator queries to get both the
runtime and registration data in a single report.

Keep the support for the old string to not break outdated clients (like
the CLAS orchestrators).