Do not check if CLARA_HOME exists on ClaraBase constructor
The environment variable CLARA_HOME only matters to the scripts that set
the CLASSPATH to run the DPEs and orchestrators, and to the DPE class to
load engines. It should not be required on construction of all CLARA
components.

Disable the null-check because otherwise even the compilation will fail
if the variable is not set and the ClaraBase constructor is called.