engine: search service info file by name first
The service info file path was constructed to follow the same package
name as the service. This is, org.jlab.clara.std.ExampleService.class
would be converted to org/jlab/clara/std/ExampleService.yaml.

But it is also useful to search just ExampleService.yaml, which is
easier to create and put in the root resources directory instead of
creating the same hierarchy as the service classfile:

    src/
        main/
          |_java/
          |   |_org/
          |       |_jlab/
          |           |_clara/
          |               |_std/
          |                   |_ ExampleService.java
          |_resources/
                ExampleService.yaml

So, let the engine specification parser search first for yaml files in
the root resources, and then try the full path with the same package as
the service.

This is in preparation to have all services described in a yaml file in
some standard location.