shell: clear interruption status before reading a new line
The commands are interrupted by catching the INT signal and interrupting
the main thread. But this means that the next iteration of the
interactive shell loop will have the interrupted status set, and the
readLine() method will be also interrupted (it uses a wait() method that
will throw, which finally results in a UserInterruptedException being
thrown). This means that after interrupting a command, there will be two
prompts, the first one being the aborted one due to the interrupted
status.

Clearing the interrupted status before reading a new line will prevent
the method throwing UserInterruptedException, and there will be just one
prompt, as expected.