shell: refactor command classes
Do not use different classes for commands and subcommands. They are the
same: have a name, a description, provide a completer and must define an
execute method that accepts the command line arguments.

Replace the current Command and Subcommand classes with a single Command
interface. The new AbstractCommand class provides a default command
implementation that requires a name and a description on the
constructor, and only needs the execute method to be defined.
For commands with subcommands, the new BaseCommand class extends the
AbstractCommand class to provide a default implementation for adding and
executing subcommands.