shell: improve reusing background local DPEs
If 'run local' is executed again, the background DPEs will be reused
instead of starting new DPEs. But if for whatever reason not all
required DPEs are still alive, then kill all the remaining and start
fresh DPEs, to prevent any issues due to corrupted DPEs and/or
registration data.