shell: workaround for C++ DPE when CTRL-C is pressed
The C++ DPE installs its own signal handler for SIGINT and SIGTERM.
So, it appears that even if Bash traps the SIGINT to do nothing, the DPE
process will still process the CTRL-C interruption.

Use the workaround from [1], this is, run the DPE in a different process
group. Now the DPE process will keep running when the user presses
CTRL-C on the shell.

TODO: there must be a better solution instead of using Perl.

[1] http://unix.stackexchange.com/a/350056