shell: pass the host and port to the local DPEs
Automatically control the address used by each DPE to support multiple
users running DPEs on the same system.

All DPEs for a run local session will share the same host address, that
can be changed from the configuration variable 'feHost'.

The port for each DPE will be chosen as the first free port from a range
of possible ports (by default ports 7000 to 8000).

Since the port of the front-end DPE will be known only at runtime,
and the orchestrator needs to know the name of the front-end,
keep the names of the DPEs along with the process instances, to get the
proper name when 'run local' is executed again.