shell: remove DPE helpers from public API
Starting a DPE by default may require more parameters for the process
builder (like setting the environment, or the working directory),
The current helpers do not support that, and changing them to run a DPE
process to access or receive all required parameters would be
cumbersome. Just make the methods private to the class that needs them.

Also, the public helpers were supposed for user-defined new commands
that may start DPEs. That seems unlikely.