shell: fix show log file command
Cannot use the local hostname to identify the log file because when
running on the farm the hostname of the farm node will be used, which is
mostly unknown to the shell that is running on some local or gateway
node.

Instead, print the last modified log file that matches the given user
and description.