shell: set all local DPE ports based on FE port
Do not use random ports for every DPE. Get a random port just for the
front-end, and then set the ports for the other DPEs based on it.

If the random port is P, then the ports will be set as follow:

    FE:     P
    Python: P + 5
    C++:    P + 10

The DPEs use a range of 3 to 5 ports. So, when searching available
ports, step 20 ports instead of 10, to reserve enough ports for
multi-lang DPEs.