shell: make the feHost variable optional
Do not set the variable by default to locahost IP. It should be set when
the user needs to use a different local IP than the resolved by CLARA.

The variable could also be used to change the IP when running on a
remote node (like the run farm command), if needed. That's why it cannot
have a local IP by default, because it will not work as the IP of the
remote node.

As it can be used to overwrite the address of both local and remote
DPEs, accept only IP addresses as values. Otherwise it gets complicated
to resolve a hostname to the proper address.