shell: make threads variable optional but with preference
Make the threads variable empty by default, which means the number of
threads will be selected automatically:

    - for local DPEs, it will be the number of cores

    - for farm DPEs, it will be the number of requested cores
      (the value of farm.cpu variable)

If the variable is set, then its value will be used for both local and
farm DPEs. This could result in the farm DPE using a different number of
cores than the requested to the farm batch system.