std: make the orchestrator wait until FE is running
Do not exit the orchestrator immediately if the front-end DPE is not
running. Keep waiting a few seconds until the DPE is alive and responds.

A common use-case is to start the DPE on the background first and then
the orchestrator. In this situation the DPE may still being initialized
when the orchestrator runs, which results in the orchestrator exiting
with an error due to not seeing the DPE alive.

To prevent the orchestrator missing the background DPE, it was required
to sleep several seconds between starting the DPE and the orchestrator,
ensuring that the background DPE is alive when the orchestrator runs.
But normally the DPE will be ready before the sleep completes, and this
makes the user wait too much.

By making the orchestrator itself wait for the DPE instead of using a
fixed sleep call, the orchestrator will continue running as soon as the
DPE is alive. Now the user doesn't need to wait all the extra seconds
until the sleep finishes to see the orchestrator running.