std: try to use local DPEs as soon as possible
When running the orchestrator, try to use the local DPEs immediately if
they are ready, instead of waiting until its alive reports are received.

The alive reports may take up to 5 seconds to arrive, due to the report
rate of the DPEs. An interactive user could see the orchestrator doing
nothing for several seconds because of this, even if he started all the
DPEs in advance.

Now if all the required local DPEs are running an registered with the
front-end, the orchestrator will start using them right after
subscribing to reconstruction DPEs. This will bypass the waiting period
for the alive reports, which may result in unnecessary delays for
interactive users.