shell: add a public builder for configuration
Clients should be able to create new configuration objects, otherwise
they cannot instantiate their commands (for example, to unit test them),
because the AbstractCommand constructor requires a Context (with a
terminal and a configuration).

When clients need to add new configuration variables, this should also
work without using the shell builder (because the tests may need the new
variables).

To make it simple to create variables in both cases (for the
client tests and also when starting the shell), use the Config builder
to add the variables and make the ClaraShell builder to set the
configuration by receiving a function that uses the Config builder.