sys: add experimental reporting to a monitor FE
If enabled, collect reports from multiple FEs and DPEs (each one
reporting by default to its own FE).

To publish reports, replace the xMsg connection with a raw ZMQ socket,
to support connecting without checking if the monitor FE is up.
The idea is that the monitor FE runs permanently in a given node, but it
may go offline for some periods.

To enabled this feature, the CLARA_MONITOR_FRONT_END environment
variable must be defined with the name of the monitor FE.

Currently, it is enabled by default for the CLAS12 farm jobs.