sys: fix stopping done/data reports
The orchestrator sends a report frequency of 0 to stop reporting
done/data messages.

But the report flag is always set to true, and the frequency (threshold)
is set to the value of 0, which means that instead of disabling the
reports, all the messages will always be reported (because the flag is
always true and the request count is always greater than the threshold).

Prevent this by setting the flag to true only if the threshold value is
greater than 0.