std: change how orchestrator checks for local DPEs
The standard orchestrator checks if the local DPEs are registered in
order to quickly start the services, instead of waiting until all the
alive reports are received. But it may be necessary to check for
registered local DPEs that use multiple local IPs (like each DPE running
in a Docker container), and the registration query can filter just by a
single IP address.

Avoid the issue by requesting all registered DPEs to the FE and then
filtering those with a local IP address (for now the same as the local
FE address).