travis: run FindBugs checks on every build
Use Travis to ensure that any warning reported by FindBugs is fixed.

Add a boolean property to set the build in CI mode, where FindBugs
warnings will be printed to console and the build will be aborted.

The FindBugs plugin only saves reports to disk, so a custom task is
needed to parse the XML report and print the warnings to console.