shell: prevent issues with JLab default terminal settings
The shell uses a wrapper script to run commands because the command has
to ignore SIGINT signals coming from the shell JVM, as introduced on
b770a3954153 ("scripts: add a wrapper that ignores SIGINT"). Otherwise a
CTRL-C on the prompt would kill the process due to signal propagation.
This is for local DPEs, which are required to keep running on the
background after finishing processing a data set.

The current implementation comes from 5dad61e45292 ("shell: improve
wrapper script to ignore CTRL-C") with a solution based on activating
Bash job control. But this makes the wrapper script to apparently not
run the command when the default JLab login script is sourced, which
sets the following terminal settings:

    stty erase "^?" kill "^U" intr "^C" eof "^D" susp "^Z" hupcl ixon ixoff tostop tabs

The actual problem is not really clear, and it is fixed when the line is
removed from the login files. But asking JLab users to edit their login
files to make the shell work is not a good user experience.

Fix this by reintroducing the first solution that uses Perl to change
the group ID of the process being executed, from 4e7ff439e4e1 ("shell:
workaround for C++ DPE when CTRL-C is pressed"), but this time check if
Perl and the required module exists. If not, use the Bash job control
solution.

As Perl is present on JLab CUE machines, the default terminal settings
should not interfere with the shell command wrappers anymore.

[skip ci]