shell: fix configuration variable builder
The expectedValues method should work with any supported variable types,
but fails when the values are not strings.

Add the missing unit tests for the builder to ensure that the bug is
fixed and that all the other methods are covered too.

Also rename the setValue(String...) method to parseValue(String...),
because when the array of values has one element, the overload method
setValues(Object) would be preferred, which is wrong because it doesn't
use the parser.