std: ensure a username when running in Docker
When running CLARA inside a Docker container, the user name may not be
defined (for example, when the --user UID option is given to docker run
but the UID is not in /etc/passwd). The user name is required to set a
default session and container names, and if it not defined it will break
CLARA: neither the session nor the container name will be valid.

Check if CLARA is running inside Docker [1] and if the user name is not
defined (the user.name property has the value '?'), and set a default
value in that case, to prevent errors.

[1]: https://stackoverflow.com/a/25518345