shell: improve quoting of external commands
Do not pass a 'quote' argument to the builder, but auto-detect if the
arguments need quotes or not.

By default only add quotes to arguments that contains spaces,
environment variables or any non-trivial characters. The other option
was to quote everything unconditionally, but that seems unnecessary, and
the output script containing the command ended up with too many quotes.
In any case, the behaviour can be toggled.

Also, when creating an array, the arguments should not be quoted
(they will be handled by the ProcessBuilder), so the 'quote' parameter
was unnecessary.