std: remove staged files from previous farm jobs
If the farm job is killed and staging is used, the staged files will be
left on disk, because the orchestrator doesn't have a shutdown hook to
remove temporary files (we cannot add a hook right now due to
limitations interrupting sync requests, that would make the hook wait a
lot of time until the request is finished).

Having files from previous jobs using space on disk may prevent the job
to run due to no space left.

As a workaround, ensure that the stage directory is empty before
starting to process the input files. But only clear the directory if the
job is using the node exclusively (this is, having required all cores in
the node).