std: send the file name to the staging service only if needed
The DataManager service will validate the 'file' parameter only if the
requested action needs it. For actions that work on the data
directories, like 'clear_stage', the 'file' parameter has no use.

This fixes ffc299ae3b17 ("std: remove staged files from previous farm
jobs"), because the input file is null when the orchestrator configures
the node, and the JSON request was sent without the 'file' parameter,
which resulted in the DataManager service returning an 'invalid request'
error.