std: fix staging service error messages
The refactoring in ce4cad6e62ca made the tests fail on JLab CUE
machines, where the /scratch directory does exist and the error message
when testing for failed staging request is not the expected.

To prevent this issue, put back a prefix for the error description
returned by the service, and test for errors by checking if the prefix
is present.

The prefix is also useful to make the error description more clear.