std: prevent files been closed twice
It may be possible to try to close the files and release the node
multiple times (see below), so to avoid closing files that were already
closed and releasing the node and setting counters more than once, check
first if the node has a current file (which is cleared after closing the
files).

This multiple attempts situation may be encountered if by luck the EOF
was finally received just when the timeout was also reached and the
timer tries to close the files.

It could also be produced if the EOF handling is updated to support
multiple EOF reports that signal to close the files (currently only the
EOF with specific severity of 2 is used for this, all the others have
severities with incremental values).