std: ignore engine errors when file are being closed
It may take a while to close the files, and if the data processing was
aborted, there will still be events being processed by the engines, and
they could report errors.

These errors should not be handled when the files are being closed.
The callback will send a request to the reader for every error, and the
reader will probably be closed already, which will result in another
error and another request, in a loop that will not finish until a new
file is opened or the orchestrator is finished. The same problem may
occur if the error is handled after the node was finished.

Use the callback critical section to prevent the errors being handled
when the node is busy closing the files, and check if the node has a
current file before restarting the chain with a new event request.