Update CheckStyle plugin to version 8.16
Sync with the version used by the IDE plugins.

Since v8.8 the indentation check has become quite annoying and flags
aligned parameters in try-with-resources as unindented, because it
results in less than 8 spaces of indentation.

The previous solution was to add suppress checks, but now the this()
constructors are also flagged, so simply reduce the line wrapping
indentation rule from 8 back to the default of 4, and the warnings go
away.

Also fix miscellaneous issues that have accumulated over time.