Configure project to use proper location for test resources
Eclipse was failing to build the project when the srcDir property for
test.resources was configured to 'test/resources', because it is as a
subdirectory of the test.java srcDir 'test', and the Eclipse "source
folders" cannot be nested.

The workaround was to make the test.resources srcDir just 'test',
and include only the resource files. But this forced to get the
resources inside the tests by prefixing them with "/resources", which
may be confusing.

It turns out it can be fixed by excluding the resources directory from
the test.java srcDir. Then the test.resources srcDir can be
'test/resources' and when getting the resource files there is no need to
use the "/resources" prefix.