Add settings.gradle file
It is recommended by the official documentation that every project
should contain a settings.gradle file [0].

Keep this file empty for now. It could be used to set the project name
as "clara", which until now has been set as the parent directory (most
probably "clara-java"). But the actual project name doesn't seem to be
important for the build or the artifacts so far. And I am not sure if it
should be set as "jclara" instead, as the artifacts are named.

[0]: https://docs.gradle.org/current/userguide/organizing_gradle_projects.html#always_define_a_settings_file