j_dpe: fix setting java.library.path
Do not set the property directly. It will override the default system
directories, and the DPE will not find native libraries installed in the
system.

Instead, add the plugin directories to the LD_LIBRARY_PATH /
DYLD_LIBRARY_PATH environment variables. These variables will be
prepended to the default java.library.path when the JVM is started.